added readme refactoring
